Moved Navigation component to its own file within layout directory
