+++ front-end/src/components/layout/Landing.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 
 export default function Landing() {
   return (
